+++ frontend/src/pages/ClubDetailPage/components/ClubDetailFooter/ClubDetailFooter.tsx
@@ -1,7 +1,7 @@
 import * as Styled from './ClubDetailFooter.styles';
 import DeadlineBadge from '@/pages/ClubDetailPage/components/DeadlineBadge/DeadlineBadge';
 import ClubApplyButton from '@/pages/ClubDetailPage/components/ClubApplyButton/ClubApplyButton';
-import { parseRecruitmentPeriod } from '@/utils/stringToDate';
+import { parseRecruitmentPeriod } from '@/utils/recruitmentPeriodParser.ts';
 import getDeadlineText from '@/utils/getDeadLineText';
 
 interface ClubDetailFooterProps {
